Use new header resolver in the service
